Reduce outgoing call jankiness
OutgoingCallBroadcaster is an activity that waits for applications like
to repond to the NEW_OUTGOING_CALL ordered broadcast.

OutgoingCallBroadcaster displays itself as a completely black activity
with a spinner (if it takes very long) until it gets a response.

When the response finally comes in, it finishes the activity and starts
the in-call UI.

Problem:
   Now that the UI is in a different process, there is a flash on the
   screen between the time the OutgoingcallBroadcaster finishes() and
   the new UI starts up

Fix:
  Introduce a 2-second delay from when we kick off the new UI until the
  OutgoingCallBroadcaster calls finish().

Change-Id: I63fcc96a7ce876ffecdf9cbad9491b502837929d