Fixing Google Voice to work with new incallui
We were not previously storing and sending the gateway information to
the UI.  This change consolidates Call gateway functionality from
InCallUiState and PhoneUtils into a new class CallGatewayManager.  It is
responsible for storing a mapping from the call's connection to the
gateway used to make that connection.  It also adds gateway packagename
and number to the Call object sent to the UI so that it can show UI
appropriately.

bug: 10393622
Change-Id: Ic5a0b068475bcab60e8cc96470273e36005ccc2e