Adding ecludeFromRecents to OutgoingCallBroadcaster
OutgoingCallBroadcaster send the Intent that triggers the call being
placed so adding excludeFromRecents in the UI layer was insufficient for
excluding the DIAL Intent from the recents bar.
The result was that Recents had an entry for incall that redialed the
call.
Fix is to add excudeFromRecents to OutgoingCallBroadcaster, which is the
source of the Intent which places the call.

bug:10412152
Change-Id: Ic1154b287352cefa3ad36ea94c0e1bd836aebc80