Adding support for dialer to bring incallUI to foreground.
Previously, the Dialer relied on
TelephonyManager.showCallScreenWithDialpad to bring the UI to the
foreground after the user selected "return to call in progress." This
was not updated for the split in-call-ui.

The fix adds a new API to CallHandlerService to bring the incallui to
the foreground.  This new API now gets called as part of
showCallScreenWithDialpad().

bug:10313347
Change-Id: Id95a1e266471970cf1295423ea720b0e1926325b