Fix DeadObjectException when bind fails.
There is a window of time during install when the service is deleted and
installed. During this time, binding to the service may cause a silent
service not found error in the log. But this error does not propagate to
our application.  Check for this case by querying the package manager
before trying to bind to the service.

Bug: 10569795
Change-Id: Ia58fde001603d0c716a2d1d91890dc31f78034a7