Do not create incoming/disconnected calls on update.
This change fixes two things:
1. onIncoming will always get called for incoming calls, but sometimes
so does onUpdate.  This caused too many updates since we ALWAYS send the
onIncoming (we need to).  The solution is to not create incoming calls
on an update any more. After this change, onUpdate will only handle
updates to incoming calls, not the creation of them.
2. Basically, the same thing happens for disconnect calls.  Some
disconnect updates come after we've disconnected and in those cases,
there is no need to recreate the objects since we've already sent a
disconnect update for the call.

bug:10571325
Change-Id: Ifa8eefd197161f90d60d4a55337a6a285e85e3bb