Cdma call waiting
There are special notifications received for call waiting on CDMA
phones.  These special notifications are handled by CallNotifier
which has special logic to handle the fact that CDMA devices do not
give us an indication of success/failure on operations to the CDMA call
waiting calls.

This fix adds functions to CallModeler that are called directly from
CallNotifier. These functions translate the events into Call objects
that it can send to the UI layer.

bug:10769004
Change-Id: Ic208ddb879c7ec76212914914b8750b55a578ac0