Cdma Outgoing 3way call support.
Cdma outgoing calls when a call already exists are special because
like normal cdma calls there is no "dialing" state.  There is also
no indication that the caller answered the phone so we have to make
guesses as to the state of the world.

This changes leverages code that already established these assumptions
can have those places call directly into CallModeler so that it can
translate those assumptions into Call objects that the UI can
understand.

Changes:
- Add concept of a generic conference call to be used for cdma
  conference calls which are very limited in nature.
- Notify CallModeler when we see an outgoing 3way call so that it
  can fake it as a "dialing" call until we are notified that it
  has "connected". This gives used the impression of a dialing state
  on cdma devices which do not have this state.
- fix CallModeler.getFullList() which previously changes model state.

bug:10769004
Change-Id: Iec5fb47fdcd1574c86120ec0fbb38911d54c7b0e