HfaActivation should support UI and no-UI modes.
Split out the HFA logic into it's own class.
Created two new classes that use the logic: HfaService and HfaActivity.

HfaService is run during setupwizard and is completely without UI.
HfaActivity is rung during activations outside of setup wizard and
uses dialogs for the UI, which provice the user the ability to skip the
provisioning if desired (just as before).

We finish the primary activity (InCallScreenShowActivation.java)
immediately when we get the request and continue the new
activity/service as a new task.

bug:10655576
Change-Id: I9b83d4253168829c82c6a1d147ac4eb25a76f39f