Update Reject with Text Message path to only use default sms app.
In KLP we have added the concept of a default SMS app to the platform. Only
this application has write permission to the SMS database. To help ensure
a better user experience we will always direct "reject via SMS" to the one
app that the user has selected as their default SMS app.

Bug: 10870624 System should direct "SENDTO" intent with "sms" and "mms" schemes to the default SMS app
Bug: 10818880 Telephony needs to direct RESPOND_VIA_MESSAGE intent only to the default SMS app
Change-Id: Id56b92285fb701db4432231950714ad2845bae60