Fix usage stats for outgoing calls.
Usage stats are updated by the CallLogger and is triggered in onDisconnect
events.  But it only updates the usage stats if a caller info has been
retrieved.  For outgoing calls, caller info retrieval was done in the
CallCard.  This no longer works since the in-call ui split.

Add caller info retrieval to placeCalls().

Bug: 10909434
Change-Id: I558a2657afd52650468da0eb4eac715364249d89