Stop sending dial intents for calls when non voice capable.
The contacts app no longer supports the dial intent since it doesn't dial.

The sole purpose of sending the dial intents was so that contacts can "add"
the number when a user somehow invokes a call intent from a tablet. But
that is a poor experience since users expect a call, not "add the number to
my contacts".

Bug: 11052202
Change-Id: I26b970b815c3ce576f28d054f35532dbe9d12196