Do not trigger RESPOND_VIA_MESSAGE with showUI Extra
We don't want to require SMS applications to implement activities that show
over lock just to handle the "respond via custom message" scenario for
rejecting phone calls. This change removes the path that would ask for
the SMS app to show UI for the user to enter a message. A separate change
adds an entry dialog into the phone call answer UI.

Bug: 11084719 Provide UI for "quick response" messages built into Phone app
Change-Id: Ibb47082313a253d506bda7978c10da13d82738a2