3way call is not working with BT
When we receive AT+CHLD=1,
we will call answerAndEndActive
which will Answers the incoming call specified by "ringing",
and ends the currently active phone call,
In this case, hanging up the active call also accepts the waiting call
in hangupActiveCall.
But ringing.isRinging() is still true,
because it take some time to change state for ringing call.
we should wait for finishing disconnecting
before check the ringing call state.
bug:10234670

Change-Id: I69031443d25303e387dd94caf79d73974e23f86a