Adds a cached phone factory.
The Phone objects being used to communicate with the radio layer listen
to the telephony layer independently to keep their own list of current
calls. Therefore, the phone object that is created to listen for
incoming calls when the device starts MUST be the same instance that the
call service uses. Additionally, since the call service itself can be
destroyed and created multiple times, we also need to reuse the phone
instance across those instantiations. This change creates a simple cache
on top of PhoneFactory to enable this.

A follow up CL will add the incoming call uses of this class.

Change-Id: I6a7bc0273101dec77908fce796b25146d95a5e6a