progress_large id defined incorrectly
Progressbar, progress_large, is defined with the wrong namespace,
removing the custom namespace will make it use the default namespace
of the application.

This affect the RRO (Runtime Resource overlay) system, making it
impossible to overlay resources in this package.
The resources which have different namespaces will be created in
different categories, the system category the first followed by
the application specific. When doing overlay the RRO will look
in the first category which will be the system, and overlay will
fail to find the resources to overlay.

Change-Id: I865cc8e85ba93b8349bef77afe41a82d8cfeea5f