Initial checkin of new Telecomm API
The "telecommframework" package will eventually be migrated to the framework
and deleted from here. It is kept here so we can quickly iterate on API changes
while implementing PSTN in terms of the new API.

At the moment, the "ConnectionService" class is complete, and the next step
is to implement "Connection" and choose one of the PSTN or SIP transports as
the initial implementation to test with.

Change-Id: Ic5d31f837f17b84419df8d469fb0e614d418b08f