Ensure Radio is on before dialing an emergency number.
Code got into a state where we were creating a Connection object to
return synchronously and then starting the Emergency Call helper after
that.  The problem is that the synchronous creation code kicked off a
"placeCall" attempt which would fail because airplane mode would be on.
We still turned on the radio but it was too late.

This change makes it so that the synchronous Connection creation is NOT
tied to the placing of a call.  Instead, we create the connection and if
it is a non-emergency call, place it, otherwise, start the emergency
call helper and wait for that to place the call.

It required that the telephony-Connection object not be set in the
constructor but instead be set through setOriginalConnection().

Bug: 16885688
Change-Id: I4acf62d34b710c2fe3439fead33b06078f5a36bf