Use description instead of generated Slot ID string.
The TelecommAccountRegistry was ignoring the display name
that the user had previously given the Settings->SIM cards activity.
It was just using a string that was generated based on the
slot id. The fix required that we listen for a content change
intent along with using the description as needed.  Note that
there is an issue (which I have filed as b/17405123). The siminfo
database seems to be regenerated on device boot.

Bug: 17390812
Change-Id: I646272add924bbe4e96e46bcf5891ed7e00e5287