Reorganize phone account settings.
+ Added new fragment for phone account settings.
+ Added new activity for phone account settings, which launches this
fragment. I would have prefered to specify a fragment on the
PreferenceScreen in call_feature_setting.xml, but I couldn't get it
to work so I dispatch an intent to start a new activity instead. I
think the preferred option doesn't work because CallFeatureSetting
is still stuck in the old PreferenceActivity model.
+ Added new activity to the manifest file.
+ Moved the default outgoing call, call assistant, and "use SIP
calling settings" into new PhoneAccountSettingsFragment.
- Deleting code for implementing those in CallFeaturesSetting.
+ Added code to AccountSelectionPreference to disable updating of the
summary, and firing and event after the value has been changed when
an account is selected.

Bug: 17487262
Change-Id: I126806ef8e70f8323a7ff143863dacffeb502eaa