Make sure that we provide the correct SIM icon based on slot ID.
Per the telephony team, you cannot calculate the slot ID from
the sub ID or the phone ID (which is what we were currently doing).
So default the slot ID to be invalid and pull the value only
if it exists in a SubInfoRecord for the SIM.  Then update the
helper function that provides the right icon based on the slot ID
to support this invalid value to provide the default icon
when the provided slot ID is not supported.

Bug: 17243556
Change-Id: I9c9faa256c0b2c8ed1121bebe909a911b9d92522