Use newly added Telecomm DisconnectCase.
+ Add utility, DisconnectCauseUtil, to convert from telephony
DisconnectCauses to telecomm DisconnectCauses. For this, map
from specific disconnect causes into a more generic set which
the UI is concerned about. Provide user-localized strings or
tones for certain cases.
- The string conversion sucks a little because it covers both
strings which are intended to be show in the InCallUi (short)
and messages which are to be displayed to the user in a dialog
when there is some termination (long). =(
+ Where possibly, use this utility for converting error
messages as well.
- Not able to convert with message and tone for many of the
failure cases, because no context exists for obtaining the
message to display.

Bug: 17329632
Change-Id: I84ee358cc8e014f110bab4c0ae7c1cb02aeebdf6