Don't call SipManager.open when creating a new SIP connection
The SIP service is already started and SipManager.open called for
all sip accounts on-device when Telecomm starts up, so there is no need
to call SipManager.open again when creating a new SIP connection.
This causes us to incorrectly open the profile without specifying that we
want to register for incoming calls.

Bug: 17305137

Change-Id: I207fe04a70212d6bdb75e7618192f16f603d9189