Include phone number in incoming call intent
This CL includes a phone number when sending an incoming
call intent. This is used by Hangouts to detect when two
incoming calls are duplicate.

This is a work around for an underlying bug where Telecomm
incorrectly shows UI for connections that are in the state
INITIALIZING.

Previouly connectinos were created asynchronously and this
allowed the connection service to do several things before
returning a initialized connection. When we moved to synchronous
creation we never updated Telecomm to handle this new state.

This underlying issue should be fixed properly for MR1.

Bug: 17485366

Change-Id: Ifa72064e0f45d799370fb72e286c509bc39f29dc