Move SIP settings to phone account settings.
+ Moved "SIP accounts" and "Receive incoming calls" settings from
the SIP settings creen to the Phone account settings screen.
+ To facilitate this, I moved some helper functions which used to
be in SipSettings to SipUtil. Their logical innards are for the
most part unchanged.
- Removed "SIP settings" preference screen added in call settings.
- Removed "Receive incoming calls" from SipSettings activity.
- Deleted unused SIP settings layout.
+ Changed back button behavior in SIP settings activity so that
back takes it to phone account settings, rather than call feature
settings.
+ Made some necessary classes/functions in sip/ public. Ihab said
that these were this way because of some refactoring at some point,
and it should be fine to make them public to achieve the necessary
class accessibility to facilitate this CL.

Bug: 17487262
Change-Id: I8e24abf624cadf9afc7aeff28781802732fef07d