Fixing where calls disconnected from conference manager are still shown.
Traced the source of this issue to the fact the calls disconnected
from the conference manager were not being recognized as disconnected
by Telecom.  When a regular call is disconnected locally, the
"precise call state changed" message from the phone reports the new
disconnected status.  However, the calls disconnected via the conference
manager were not reporting their disconnected status via the "precise call
state changed handler", and thus they were not recognized as disconnected.

To fix this, in TelephonyConnections, registered a disconnect handler
to listen to disconnections from the Phone.  This fixed the problem.

Bug: 17785678
Change-Id: I8b0f94f3afd2d8a961d75f7e309c598934e405c3