Handle out of index sip call options.
If the sip call option reported by sip shared preferences is not
found, force it to be SIP_ADDRESS_ONLY as a default.

I'm inferring that the trace in this bug happened because of an
invalid value set for the sip call option, so that it isn't
found as one of the possible values.

I'm supposing this could happen if a phone had the deprecated
SIP_ASK_ME_EACH_TIME value set (which is no longer a valid option).

Not positive if it's what's happening in this bug, but it seemed
like a reasonable case to safeguard against.

Bug: 17885832
Change-Id: I6d66ae3c4de7ccea38de3959a1f84cdfe2e114e1