Cleanup Call Settings initialization code.
Most of these null checks shouldn't be necessary, because we're
adding the preferences every time onResume is called and updating.
That is; from what I understand we expect these buttons should
always exist, and if not we:would want to know.

I don't understand why updating the settings was deferred to later
in onResume and also protected by null checks. I figure it can be
done at the same time.

Motivation for this is that I need to pull out some call forwarding
settings into a more generic form, and trying to clean up the
Call Settings while I'm at it to make the logic of it all simpler,
or, at the very least, more concise.

Bug: 17947839
Change-Id: I3a61eff5fd8b42de3d7e06c8bf9cbc33326397b9