Disable disconnect/separate for connections which were ever IMS.
When a conference call hands over from an IMS call to a circuit
switch call, we aren't guaranteed to support a call list and be
able to actions like disconnect/separate which you might if you
started the conference in GSM.

For now, disable DISCONNECT/SEPARATE if the connection is was
ever an IMS connection. We can adjust casing depending on
the individual carrier IMS implementations later on.

Also, if a conference is handed over from IMS to CS, disable
conference management because we don't know if we'll be able to
regenerate the call list.

If we need to track more initial states in the future, we should
use an enum or set of integer states instead of a boolean to track.

Bug: 17429707
Change-Id: Ib23977c8f731395bb70628a2a7af5b4052808d49