Set ADD_CALL capabilities on TelephonyConnections.
Before this was done purely in Telecom's InCallController. But, IMS
now adds some wrinkles on the situations where this is possible. Now,
specify this on the connection, and remove on the InCallController if
the capability is not appropriate.

This allows for greater control over ADD_CALL capability, depending
on the type/state of the TelephonyConnection.

Bug: 17423848
Change-Id: I629b87e1652ddfd7008f56e9b46669b9343232db