Remove manage conference option for IMS.
Before, this was only done for CDMA, but upon thought it should be
disabled in general, unless there's a conference event package.
I figure this should be included as default behavior, rather than
assuming that Gsm will have a conference event package, as is the
case right now because it hasn't yet been implemented.

Bug: 17423848
Change-Id: I573460b72c6cd8eca4a5618d74a2dab79caa897b