Clean up some voicemail methods to be more clear.
- Function which handled voicemail button click request was short
and only called in one place, so I moved the code to where it is
invoked.

+ Added constant for magical int and rewrote isUpdateRequired to
express the case it's checking for more clearly.

- Remove overriden method for startActivityForResult(), which
doesn't seem to apparently add anything besides a debug statement..

+ This is somewhat of a subjective call, but I found the way in
which some of the checks for success/failure were confusing, in
part because debug/log statements were interleaved with the code.
I reworked these so that the return value/logging are no longer
dependent on each other.

Bug: 17019623
Change-Id: Id9bbe17a51213a4c1084af694d0d4becc0edea56