Add voicemail settings activity and fragment.
These are just stubs for now; in subsequent CLs, I will move the
settings/logic from CallFeatureSettings to these files.

Removed the comment in PhoneAccountSettingsActivity; the current
vision is that the activity will remain. This is particularly
useful because on Multi-SIM phones, we will want to be able to
launch phone accounts directly with an intent.

Bug: 17019623
Change-Id: Ie3f4aeca848f348e0e38af3dbd8bce0bd9f97667