Rewrite voicemail initialization code.
All this index trickery with the default voicemail provider wasn't
intuitive to me until going through the code in more detail.

I figured to make this list-based instead of array-based. It might
be marginally worse in terms of performance (although it does save
on a loop), but I think results in more readable logic.

Bug: 17019623
Change-Id: I153fcc9e0f9bbd65a71a89cf355313cdd0609bf6