Add connections to conference controllers when original connection is valid
The only way that we can determine if we should use the Telephony or CDMA
conference controllers are once the underlying original connection is set.
Until it is set, we cannot guarantee that we'll get it right. This CL introduces
some plumbing that allows the TelephonyConnection to notify the
TelephonyConnectionService when this occurs so that it can take action
at the appropriate time.

Bug: 18240234
Bug: 18056632
Change-Id: I54b4d9821b94085e639c35f2fe3569b4eed67f88