Add utility for voicemail dialogs.
Try to isolate voicemail dialog-related constructs. I'm not sure if
this is as sensible as some of the other refactoring work I've done,
but the reasoning behind it is...

+ The logic is self-standing.
+ It incrementally moves a chunk of the voicemail code out of call
settings to help reduce the churn when a lot of other more connected
pieces have to be moved.
+ It provides greater isolation to the dialog logic, which I think
may be handy as a step since it seems that the VoicemailFragmen
 will need to contain a DialogFragment, similarly to how the
SettingsPreferenceFragment works.

Bug: 18114923
Change-Id: I0680869bf2d3e5b030cbcfd7df17bff46912f50f