Create IMSConference to represent IMS conference calls.
- ImsConference contains a "host" connection which is the radio connection
to the IMS conference server.
When a new ImsConference is created, we wrap the radio connection in a
new TelephonyConnection which is not added to Telecom; this is stored in
the ImsConference and provides a means for the conference to be notified
of disconnect/hold of the conference connection.
Next, the original TelephonyConnection is disconnected and the conference
is added to Telecom.
- The ImsConference manages its own conference participants directly
rather than through the ImsConference controller.
- Added a new ImsConferenceController since it is sufficiently different
from how the generic TelephonyConferenceController works.

Bug: 18200934
Change-Id: Ia32c2111173c69be7b05ed76a038fb571c2e726b