Fixing issue where DTMF tones were not issued through to the IMS framework once a conference is established.
The problem was the "ConferenceHostConnection" which wrapped the radio
connection.  As an extension of TelephonyConnection, it did not provide
the required "onPlayDtmfTone" and "onStopDtmfTone" overrides the CDMA
and GSM TelephonyConnections do.

Refactored to get rid of the ConferenceHostConnection and just clone the
original telephony connection instead.  This ensures the correct DTMF
overrides are present.

Bug: 18200934
Change-Id: Iff8d404ff494be39ef68ba10b23c3a38b74e520d