SRVCC Conference call scenarios with mid-call enabled.
1. Disconnect all IMS conference participants on Handover. Create a
telephony connection out of the original connection part of
IMS conference host and add that to Connection Service as an
existing connection. Add the new Telephony connection to the
conference controller.

Telephony Conference controller delays the attempt to add the
conference to the connection service, until all the participants
are already part of the ConnectionService.

Retrigger the conference recalculation of conference controller
after a Unknown connection is added to ConnectionService.

2. Add support for SRVCC conference with greater than 3 participants

After the Gsm Conference object is created with the first two
participants, delay adding the subsequent unknown connections to
the conference until the connections are added to connection
service.

3. Add support for using the CS connection address during SRVCC

The check for pre-handover state is only applicable to IMS
connection. Compare the state of new handover connection with
pre-handover state of Ims Connection to process the correct update
corresponding to the TelephonyConnection.

Change-Id: I2ab2ee02a83bf8555856ccdba3de0aae75041f9b