Listen to subscription info changes to update MWI.
MWI refers to "message waiting indicator" (aka. voicemail
notifications). Before, there custom delay/retry code to handle if
the SIM was not yet registered, but the network indicated a voicemail.

This CL deletes that custom logic, and instead listens for changes
to the subscription info (which should change with SIM registration)
as a trigger for retrying to show the message waiting. I moved some
logic into NotificationMgr so this behavior can be centralized in
one class.

This clears the way to implement some MSIM voicemail notification
settings; the retry logic was written assuming there was only one
internal.telephony.Phone, but there may now be multiple ones. This
change makes it easier to iterate over all the phones and update
their MWIs, which will be done in a subsequent CL.

It's a little tricky to test the exact scenario which this case
accounted for; I tested by removing/re-inserting a SIM to see if
it updated accordingly (which it did successfully).

Bug: 18232725
Change-Id: Ica2c5a2ddef8a2efc2c40c2f4d8729b485eb0bf7