Create VoicemailProviderListPreference.
Move logic specific to the preference which is a list of
voicemail providers into its own class. This will provide a clearer
separation for adapting these settings to be keyed by subscription id
and handle any associated migration.

Right now the default preference behavior, using the name of the
preference, is used to persist the setting. I plan to change this so
that we can provide a key which is subscription-sensitive.

Bug: 18232725
Bug: 17019623
Change-Id: I5e7d53b048ef57d84dde1bdde5e35a9cb55a7bbb