Derive "merged" subscribers based on line numbers.
Adds an API to indicate a set of subscriber identities that should
be considered as "merged" together for data usage purposes.  The
internal implementation clusters identities using the carrier-provided
line number override as a key.

Currently only supports a single merge set, which is selected
arbitrarily based on the first active subscriber with a line number
override.

Bug: 18012787
Change-Id: I9a0b1d1850bf1cd1fe9fe90984b00600f7d1c689